tests: add disconnect tests
